change kill logic
